requirement.txt
apache-airflow-providers-postgres
apache-airflow-providers-pgvector
langchain-openai
langchain-community
pypdf